+++ src/shogun/mathematics/graph/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(SHOGUN_CORE_GRAPH_HEADERS
 	${CMAKE_CURRENT_SOURCE_DIR}/Array.h
 	${CMAKE_CURRENT_SOURCE_DIR}/Graph.h
+	${CMAKE_CURRENT_SOURCE_DIR}/GraphEnv.h
 	${CMAKE_CURRENT_SOURCE_DIR}/GraphExecutor.h
 	${CMAKE_CURRENT_SOURCE_DIR}/Shape.h
 	${CMAKE_CURRENT_SOURCE_DIR}/Tensor.h
@@ -23,6 +24,7 @@ set(SHOGUN_CORE_GRAPH_HEADERS
 
 set(SHOGUN_CORE_GRAPH_SOURCES
 	${CMAKE_CURRENT_SOURCE_DIR}/Graph.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/GraphEnv.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/GraphExecutor.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/Tensor.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/Types.cpp
